refactor(dynamodb): simplify code
